+++ index.common.js
@@ -13,7 +13,4 @@ exports.isValidPhoneNumber    = require('./build/phone').validate
 exports.format_phone_number = require('./build/phone').format
 exports.formatPhoneNumber   = require('./build/phone').format
 
-exports.format_phone_number_international = require('./build/phone').format_international
-exports.formatPhoneNumberInternational    = require('./build/phone').format_international
-
 exports['default'] = Input