+++ index.es6.js
@@ -7,8 +7,6 @@ export
 	validate as is_valid_phone_number,
 	validate as isValidPhoneNumber,
 	format as format_phone_number,
-	format as formatPhoneNumber,
-	format_international as format_phone_number_international,
-	format_international as formatPhoneNumberInternational
+	format as formatPhoneNumber
 }
 from './source/phone'